+++ Projects/FeatureMemo/Targets/FeatureMemo/Memo.swift
@@ -1,0 +1,9 @@
+//
+//  Memo.swift
+//  FeatureMemo
+//
+//  Created by 송영모 on 2023/03/17.
+//  Copyright © 2023 Copynote. All rights reserved.
+//
+
+import Foundation
