+++ Projects/FeatureMemo/Targets/FeatureMemo/MemoView.swift
@@ -1,0 +1,17 @@
+//
+//  MemoView.swift
+//  FeatureMemo
+//
+//  Created by 송영모 on 2023/03/17.
+//  Copyright © 2023 Copynote. All rights reserved.
+//
+
+import SwiftUI
+
+struct MemoView: View {
+    var body: some View {
+        List {
+            
+        }
+    }
+}
